dtype

+++ lib/networks/latent_xyzc.py
@@ -61,7 +61,7 @@ class Network(nn.Module):
 
     def interpolate_features(self, grid_coords, feature_volume):
         features = []
-        for volume in feature_volume:
+        for volume in feature_volume.double():
             feature = F.grid_sample(volume,
                                     grid_coords,
                                     padding_mode='zeros'
